Created a new table 'likes'.

+++ assets/02_create_tb_posts.sql
@@ -1,6 +1,6 @@
 CREATE TABLE posts (
-  post_id INT(3) NOT NULL AUTO_INCREMENT PRIMARY KEY
-  ,post_category_id INT(3) NOT NULL
+  post_id TINYINT(3) NOT NULL AUTO_INCREMENT PRIMARY KEY
+  ,post_category_id TINYINT(3) NOT NULL
   ,post_title VARCHAR(255) NOT NULL
   ,post_author VARCHAR(255) NOT NULL
   ,post_date DATETIME NOT NULL
